View one recipe

+++ backend/src/main/kotlin/com/agagnier/menugenerator/backend/dto/IngredientDto.kt
@@ -2,6 +2,6 @@ package com.agagnier.menugenerator.backend.dto
 
 import com.agagnier.menugenerator.model.RecipeIngredient
 
-data class IngredientDto(val ingredientId: Int, val name: String, val quantity: Int, val unit: String, val section: String?)
+data class IngredientDto(val ingredientId: Int, val name: String, val quantity: Float, val unit: String, val section: String?)
 
 fun toIngredientDto(i: RecipeIngredient) = IngredientDto(i.ingredient.id.value, i.ingredient.name, i.quantity, i.unit, i.section)
